frs-backend/requirements.txt
anyio==3.6.1
asgiref==3.5.2
autopep8==1.7.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
cmake==3.24.1.1
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
Django==4.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
dlib==19.24.0
drf-yasg==1.21.3
face-recognition==1.3.0
face-recognition-models==0.3.0
fastapi==0.82.0
h11==0.13.0
httptools==0.4.0
idna==3.3
inflection==0.5.1
install==1.3.5
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.2
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.2.0
pycodestyle==2.9.1
pydantic==1.10.2
PyJWT==2.4.0
pyparsing==3.0.9
pytesseract==0.3.10
python-dotenv==0.21.0
python-slugify==6.1.2
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sniffio==1.3.0
sqlparse==0.4.2
starlette==0.19.1
text-unidecode==1.3
toml==0.10.2
typing-extensions==4.3.0
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.12
uvicorn==0.18.3
watchfiles==0.16.1
websockets==10.3
faker==14.2.0